Move to correct module -> eliminate dependency
